Shopped cover image
Used Photoshop to add the overlay to the cover image directly instead
of having to mess around with weird sizing of the overlay to fit the
entire window.

Removed extraneous CSS rules from previous implementations.

Renamed the ID’s of the new implementation stuff to the appropriate
names instead of test names.